+++ spec/mongoid/relations/embedded/many_spec.rb
@@ -2549,7 +2549,7 @@ describe Mongoid::Relations::Embedded::Many do
     it "returns the valid options" do
       described_class.valid_options.should eq(
         [
-          :as, :cascade_callbacks, :cyclic, :order, :versioned,
+          :as, :cascade_callbacks, :cyclic, :order,
           :store_as, :before_add, :after_add, :before_remove, :after_remove
         ]
       )